remove 2019 and 2024 Uk Parliament source data parsing duplication
+ moved resources directory lookups to Config class
