+++ src/main/java/com/knocksfornometer/mapimage/config/Config.java
@@ -9,18 +9,22 @@ import java.net.URI;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.knocksfornometer.mapimage.Main.CONFIG;
+
 public class Config {
 
+    public static final String CONFIG_JSON_PATH = "target/classes/mapimage-config.json";
     public Map<ElectionYearDataSource, ElectionDataSourceConfig> electionDataSources = new HashMap<>();
 
     public String targetOutputBaseDir;
     public String resourcesDir;
+    public String svgMapInputFileNameSuffix;
 
     @SneakyThrows
     public static Config load(){
         var objectMapper = new ObjectMapper();
         objectMapper.configure(JsonParser.Feature.ALLOW_COMMENTS, true);
-        return objectMapper.readValue(new URI("file:target/classes/mapimage-config.json").toURL(), Config.class);
+        return objectMapper.readValue(new URI("file:" + CONFIG_JSON_PATH).toURL(), Config.class);
     }
 
     public ElectionDataSourceConfig get(final ElectionYearDataSource electionDataSource){
@@ -29,5 +33,16 @@ public class Config {
 
     public static class ElectionDataSourceConfig {
 
+    }
+
+    public String getSvgMapInputFileName(final int year) {
+        return year + svgMapInputFileNameSuffix;
+    }
+
+    public String getElectionYearDataSourceResourcePath(final ElectionYearDataSource electionYearDataSource){
+        return CONFIG.resourcesDir +
+                "\\election_data\\%d\\%s\\".formatted(
+                        electionYearDataSource.getElectionYear().getYear(),
+                        electionYearDataSource.getElectionDataSource().name().toLowerCase());
     }
 }
