+++ src/main/java/com/knocksfornometer/mapimage/data/ElectionDataSource.java
@@ -1,3 +1,3 @@
 package com.knocksfornometer.mapimage.data;
 
-public enum ElectionDataSource { ElectoralCommission, UkParliament }
+public enum ElectionDataSource {ELECTORAL_COMMISSION, UK_PARLIAMENT}
