+++ src/main/java/com/knocksfornometer/mapimage/domain/ElectionData.java
@@ -4,13 +4,9 @@ import java.util.Map;
 
 import com.knocksfornometer.mapimage.data.ElectionYearDataSource;
 
-/**
- * @author Eelco de Vlieger
- */
 public record ElectionData(ElectionYearDataSource electionYearDataSource,
 						   Map<String, String> partyColorMapping,
 						   String svgMapInputFile,
 						   ConstituencyKeyGenerator constituencyKeyGenerator) {
-
 
 }
