+++ src/main/java/com/knocksfornometer/mapimage/imagegeneration/ImageGenerator.java
@@ -3,8 +3,6 @@ import java.awt.image.WritableRaster;
 
 /**
  * Generate an image using a {@link WritableRaster}
- * 
- * @author Eelco de Vlieger
  */
 public interface ImageGenerator{
 	/** Generate an image using a {@link WritableRaster} */
